fix(core): apply session-read guard to all call sites + add tests
Addresses review feedback on #1538 — the guard was under-applied. Two
structurally identical bare `await session.get("user")` reads were left
unhardened:

  - middleware.ts: the main middleware runs *before* the auth middleware
    (both order: "pre"), so on a session-bearing request this is the *first*
    session read — it could hang and poison the isolate before the auth fix
    ever runs.
  - routes/api/snapshot.ts: a PUBLIC_API_EXACT route the auth middleware skips,
    so its manual read is the only session lookup.

Extract the guard to a shared `astro/session-user.ts` (`resolveSessionUser`)
and apply it at all three sites (main middleware keeps its `!hasSessionCookie`
short-circuit, the #733 KV-read-miss optimization). Add unit tests for the two
testable safety behaviors: timeout backstop on a never-settling read, and
fail-closed on a rejecting read.
